Aperfeiçoando os módulos vuex
