+++ gabriel-alves-dos-santos/e-commerce-challenge/src/components/CartComponent.vue
@@ -75,7 +75,7 @@ export default {
     const store = useStore()
     const { CLEAN_LIST } = types
     
-    state.cartData = computed(() => store.state.listaDeCompras)
+    state.cartData = computed(() => store.state.cart.listaDeCompras)
     state.totalPrice = computed(() => store.getters.getTotalPrice)
     state.totalItens = computed(() => store.getters.getTotalItens)
