+++ gabriel-alves-dos-santos/e-commerce-challenge/src/components/DataContainer.vue
@@ -35,7 +35,7 @@ export default {
     })
     const store = useStore()
     const { CLEAN_DATA, ADD_TO_CART } = types
-    state.data = computed(() => store.state.dataContainer)
+    state.data = computed(() => store.state.dados.dataContainer)
 
     
     function turnToPrice(num){
@@ -80,8 +80,8 @@ export default {
         }
       }  
     )
-    watch(() => store.state.selectedCategory,
-        () => getCategory(store.state.selectedCategory),
+    watch(() => store.state.dados.selectedCategory,
+        () => getCategory(store.state.dados.selectedCategory),
         {
           deep: true
         }
